core_backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
bcrypt==4.1.2
pytest==8.3.1
python-dotenv==1.0.1